+++ framework/contracts/account/manager/tests/proxy.rs
@@ -1,5 +1,6 @@
 mod common;
-use abstract_adapter::mock::MockExecMsg;
+use abstract_adapter::mock::{MockExecMsg, MockInitMsg};
+use abstract_core::adapter::{AdapterBaseMsg, AdapterRequestMsg};
 use abstract_core::manager::{ModuleInstallConfig, ModuleVersionsResponse};
 use abstract_core::objects::fee::FixedFee;
 use abstract_core::objects::module::{ModuleInfo, ModuleVersion, Monetization};
@@ -106,7 +107,7 @@ fn default_without_response_data() -> AResult {
 
     let resp = account.manager.execute_on_module(
         TEST_MODULE_ID,
-        Into::<abstract_core::adapter::ExecuteMsg<MockExecMsg>>::into(MockExecMsg),
+        Into::<abstract_core::adapter::ExecuteMsg<MockExecMsg>>::into(MockExecMsg {}),
     )?;
     assert_that!(resp.data).is_none();
     take_storage_snapshot!(chain, "default_without_response_data");
@@ -156,11 +157,7 @@ fn install_standalone_modules() -> AResult {
         ModuleReference::Standalone(standalone1_id),
     )])?;
 
-    account.install_module(
-        "abstract:standalone1",
-        Some(&mock_modules::standalone_cw2::MockMsg),
-        None,
-    )?;
+    account.install_module("abstract:standalone1", Some(&MockInitMsg {}), None)?;
 
     // install second standalone
     deployment.version_control.propose_modules(vec![(
@@ -172,11 +169,7 @@ fn install_standalone_modules() -> AResult {
         ModuleReference::Standalone(standalone2_id),
     )])?;
 
-    account.install_module(
-        "abstract:standalone2",
-        Some(&mock_modules::standalone_no_cw2::MockMsg),
-        None,
-    )?;
+    account.install_module("abstract:standalone2", Some(&MockInitMsg {}), None)?;
     take_storage_snapshot!(chain, "proxy_install_standalone_modules");
     Ok(())
 }
@@ -206,11 +199,7 @@ fn install_standalone_versions_not_met() -> AResult {
     )])?;
 
     let err = account
-        .install_module(
-            "abstract:standalone1",
-            Some(&mock_modules::standalone_cw2::MockMsg),
-            None,
-        )
+        .install_module("abstract:standalone1", Some(&MockInitMsg {}), None)
         .unwrap_err();
 
     if let AbstractInterfaceError::Orch(err) = err {
@@ -302,11 +291,11 @@ fn install_multiple_modules() -> AResult {
             vec![
                 ModuleInstallConfig::new(
                     ModuleInfo::from_id_latest("abstract:standalone1")?,
-                    Some(to_json_binary(&mock_modules::standalone_cw2::MockMsg).unwrap()),
+                    Some(to_json_binary(&MockInitMsg {}).unwrap()),
                 ),
                 ModuleInstallConfig::new(
                     ModuleInfo::from_id_latest("abstract:standalone2")?,
-                    Some(to_json_binary(&mock_modules::standalone_no_cw2::MockMsg).unwrap()),
+                    Some(to_json_binary(&MockInitMsg {}).unwrap()),
                 ),
             ],
             Some(&[coin(86, "token1"), coin(500, "token2")]),
@@ -322,11 +311,11 @@ fn install_multiple_modules() -> AResult {
     account.install_modules_auto(vec![
         ModuleInstallConfig::new(
             ModuleInfo::from_id_latest("abstract:standalone1")?,
-            Some(to_json_binary(&mock_modules::standalone_cw2::MockMsg).unwrap()),
+            Some(to_json_binary(&MockInitMsg {}).unwrap()),
         ),
         ModuleInstallConfig::new(
             ModuleInfo::from_id_latest("abstract:standalone2")?,
-            Some(to_json_binary(&mock_modules::standalone_no_cw2::MockMsg).unwrap()),
+            Some(to_json_binary(&MockInitMsg {}).unwrap()),
         ),
     ])?;