+++ framework/contracts/native/account-factory/tests/create_account.rs
@@ -14,6 +14,7 @@ use abstract_core::{
 use abstract_interface::{
     AbstractAccount, AccountFactoryExecFns, AccountFactoryQueryFns, VCQueryFns, *,
 };
+use abstract_sdk::cw_helpers::Clearable;
 use abstract_testing::prelude::*;
 use cosmwasm_std::Addr;
 use cw_asset::{AssetInfo, AssetInfoBase};
@@ -82,7 +83,9 @@ fn create_one_account() -> AResult {
 
     let vc_config = version_control.config()?;
     let expected = abstract_core::version_control::ConfigResponse {
-        factory: factory.address()?,
+        account_factory_address: Some(factory.address()?),
+        allow_direct_module_registration_and_updates: true,
+        namespace_registration_fee: Default::default(),
     };
 
     assert_that!(&vc_config).is_equal_to(&expected);
@@ -156,7 +159,9 @@ fn create_two_account_s() -> AResult {
 
     let vc_config = version_control.config()?;
     let expected = abstract_core::version_control::ConfigResponse {
-        factory: factory.address()?,
+        account_factory_address: Some(factory.address()?),
+        allow_direct_module_registration_and_updates: true,
+        namespace_registration_fee: Default::default(),
     };
 
     assert_that!(&vc_config).is_equal_to(&expected);