+++ framework/packages/abstract-adapter/src/endpoints/execute.rs
@@ -8,16 +8,13 @@ use abstract_core::adapter::AdapterBaseMsg;
 use abstract_core::adapter::{AdapterExecuteMsg, AdapterRequestMsg, BaseExecuteMsg, ExecuteMsg};
 use abstract_core::manager::state::ACCOUNT_MODULES;
 use abstract_core::objects::nested_admin::query_top_level_owner;
-use abstract_sdk::AccountAction;
+
 use abstract_sdk::{
     base::{ExecuteEndpoint, Handler, IbcCallbackEndpoint, ReceiveEndpoint},
     features::ModuleIdentification,
-    AbstractResponse, AccountVerification, Execution, ModuleInterface,
+    AbstractResponse, AccountVerification,
 };
-use cosmwasm_std::{
-    wasm_execute, Addr, CosmosMsg, Deps, DepsMut, Env, MessageInfo, QuerierWrapper, Response,
-    StdError, StdResult,
-};
+use cosmwasm_std::{Addr, Deps, DepsMut, Env, MessageInfo, QuerierWrapper, Response, StdResult};
 use schemars::JsonSchema;
 use serde::Serialize;
 
@@ -47,8 +44,6 @@ impl<
                 .map_err(From::from),
             ExecuteMsg::IbcCallback(msg) => self.ibc_callback(deps, env, info, msg),
             ExecuteMsg::Receive(msg) => self.receive(deps, env, info, msg),
-            #[allow(unreachable_patterns)]
-            _ => Err(StdError::generic_err("Unsupported api execute message variant").into()),
         }
     }
 }
@@ -65,7 +60,7 @@ impl<Error: ContractError, CustomInitMsg, CustomExecMsg, CustomQueryMsg, Receive
     fn base_execute(
         &mut self,
         deps: DepsMut,
-        env: Env,
+        _env: Env,
         info: MessageInfo,
         message: BaseExecuteMsg,
     ) -> AdapterResult {
@@ -105,7 +100,6 @@ impl<Error: ContractError, CustomInitMsg, CustomExecMsg, CustomQueryMsg, Receive
             AdapterBaseMsg::UpdateAuthorizedAddresses { to_add, to_remove } => {
                 self.update_authorized_addresses(deps, info, to_add, to_remove)
             }
-            AdapterBaseMsg::Remove {} => self.remove_self_from_deps(deps.as_ref(), env),
         }
     }
 
@@ -163,42 +157,6 @@ impl<Error: ContractError, CustomInitMsg, CustomExecMsg, CustomQueryMsg, Receive
         };
         self.target_account = Some(account_base);
         self.execute_handler()?(deps, env, info, self, request.request)
-    }
-
-    /// If dependencies are set, remove self from them.
-    fn remove_self_from_deps(&mut self, deps: Deps, env: Env) -> AdapterResult {
-        let account_base = self.target_account.as_ref().unwrap();
-
-        let dependencies = self.dependencies();
-        let mut msgs: Vec<CosmosMsg> = vec![];
-        let modules = self.modules(deps);
-        for dep in dependencies {
-            let adapter_addr = modules.module_address(dep.id);
-            // just skip if dep is already removed. This means all the authorized addresses are already removed.
-            if adapter_addr.is_err() {
-                continue;
-            };
-            msgs.push(
-                wasm_execute(
-                    adapter_addr?.into_string(),
-                    &BaseExecuteMsg {
-                        msg: AdapterBaseMsg::UpdateAuthorizedAddresses {
-                            to_add: vec![],
-                            to_remove: vec![env.contract.address.to_string()],
-                        },
-                        proxy_address: Some(account_base.proxy.to_string()),
-                    },
-                    vec![],
-                )?
-                .into(),
-            );
-        }
-        self.executor(deps)
-            .execute_with_response(
-                vec![AccountAction::from_vec(msgs)],
-                "remove_adapter_from_dependencies",
-            )
-            .map_err(Into::into)
     }
 
     /// Update authorized addresses from the adapter.
@@ -531,7 +489,7 @@ mod tests {
 
             let msg = ExecuteMsg::Module(AdapterRequestMsg {
                 proxy_address: None,
-                request: MockExecMsg,
+                request: MockExecMsg {},
             });
 
             let unauthorized: String = "someoone".into();
@@ -561,7 +519,7 @@ mod tests {
 
             let msg = ExecuteMsg::Module(AdapterRequestMsg {
                 proxy_address: None,
-                request: MockExecMsg,
+                request: MockExecMsg {},
             });
 
             let res = execute_as(deps.as_mut(), TEST_MANAGER, msg);
@@ -578,7 +536,7 @@ mod tests {
 
             let msg = ExecuteMsg::Module(AdapterRequestMsg {
                 proxy_address: None,
-                request: MockExecMsg,
+                request: MockExecMsg {},
             });
 
             let res = execute_as(deps.as_mut(), TEST_AUTHORIZED_ADDRESS, msg);
@@ -595,7 +553,7 @@ mod tests {
 
             let msg = ExecuteMsg::Module(AdapterRequestMsg {
                 proxy_address: Some(TEST_PROXY.into()),
-                request: MockExecMsg,
+                request: MockExecMsg {},
             });
 
             let res = execute_as(deps.as_mut(), TEST_AUTHORIZED_ADDRESS, msg);
@@ -619,7 +577,7 @@ mod tests {
 
             let msg = ExecuteMsg::Module(AdapterRequestMsg {
                 proxy_address: Some(other_proxy.into()),
-                request: MockExecMsg,
+                request: MockExecMsg {},
             });
 
             let res = execute_as(deps.as_mut(), TEST_AUTHORIZED_ADDRESS, msg);