+++ framework/packages/abstract-adapter/src/lib.rs
@@ -58,19 +58,19 @@ pub mod mock {
     }
 
     #[cosmwasm_schema::cw_serde]
-    pub struct MockInitMsg;
+    pub struct MockInitMsg {}
 
     #[cosmwasm_schema::cw_serde]
-    pub struct MockExecMsg;
+    pub struct MockExecMsg {}
 
     #[cosmwasm_schema::cw_serde]
-    pub struct MockQueryMsg;
+    pub struct MockQueryMsg {}
 
     #[cosmwasm_schema::cw_serde]
-    pub struct MockReceiveMsg;
+    pub struct MockReceiveMsg {}
 
     #[cosmwasm_schema::cw_serde]
-    pub struct MockSudoMsg;
+    pub struct MockSudoMsg {}
 
     /// Mock Adapter type
     pub type MockAdapterContract = AdapterContract<
@@ -111,7 +111,7 @@ pub mod mock {
                 ans_host_address: TEST_ANS_HOST.into(),
                 version_control_address: TEST_VERSION_CONTROL.into(),
             },
-            module: MockInitMsg,
+            module: MockInitMsg {},
         };
         adapter.instantiate(deps, mock_env(), info, init_msg)
     }
@@ -126,7 +126,7 @@ pub mod mock {
                 ans_host_address: TEST_ANS_HOST.into(),
                 version_control_address: TEST_VERSION_CONTROL.into(),
             },
-            module: MockInitMsg,
+            module: MockInitMsg {},
         };
         adapter.instantiate(deps, mock_env(), info, init_msg)
     }
@@ -207,6 +207,88 @@ pub mod mock {
         impl <T: ::cw_orch::prelude::CwEnv> ::abstract_interface::AdapterDeployer<T, MockInitMsg> for $name <T> {}
 
         impl<T: ::cw_orch::prelude::CwEnv> Uploadable for $name<T> {
+            fn wrapper(&self) -> <Mock as ::cw_orch::environment::TxHandler>::ContractSource {
+                Box::new(ContractWrapper::<
+                    Exec,
+                    _,
+                    _,
+                    _,
+                    _,
+                    _,
+                >::new_with_empty(
+                    self::execute,
+                    self::instantiate,
+                    self::query,
+                ))
+            }
+        }
+
+        impl<Chain: ::cw_orch::environment::CwEnv> $name <Chain> {
+            pub fn new_test(chain: Chain) -> Self {
+                Self(
+                    ::cw_orch::contract::Contract::new($id, chain),
+                )
+            }
+        }
+    };
+}
+
+    /// Generate a BOOT instance for a 0.19 abstract mock adapter
+    /// - $name: name of the contract (&str)
+    /// - $id: id of the contract (&str)
+    /// - $version: version of the contract (&str)
+    /// - $deps: dependencies of the contract (&[StaticDependency])
+    #[macro_export]
+    macro_rules! gen_adapter_old_mock {
+    ($name:ident, $id:expr, $version:expr, $deps:expr) => {
+        use ::abstract_core::adapter::*;
+        use ::cosmwasm_std::Empty;
+        use ::abstract_adapter::mock::{MockExecMsg, MockQueryMsg, MockReceiveMsg, MockInitMsg, MockAdapterContract, MockError};
+        use ::cw_orch::environment::CwEnv;
+
+        const MOCK_ADAPTER: ::abstract_adapter::mock::MockAdapterContract = ::abstract_adapter::mock::MockAdapterContract::new($id, $version, None)
+        .with_dependencies($deps);
+
+        fn instantiate(
+            deps: ::cosmwasm_std::DepsMut,
+            env: ::cosmwasm_std::Env,
+            info: ::cosmwasm_std::MessageInfo,
+            msg: <::abstract_adapter::mock::MockAdapterContract as ::abstract_sdk::base::InstantiateEndpoint>::InstantiateMsg,
+        ) -> Result<::cosmwasm_std::Response, <::abstract_adapter::mock::MockAdapterContract as ::abstract_sdk::base::Handler>::Error> {
+            use ::abstract_sdk::base::InstantiateEndpoint;
+            MOCK_ADAPTER.instantiate(deps, env, info, msg)
+        }
+
+        /// Execute entrypoint
+        fn execute(
+            deps: ::cosmwasm_std::DepsMut,
+            env: ::cosmwasm_std::Env,
+            info: ::cosmwasm_std::MessageInfo,
+            msg: ::abstract_core::base::ExecuteMsg<::abstract_core::adapter::AdapterBaseMsg, MockExecMsg, MockReceiveMsg>,
+        ) -> Result<::cosmwasm_std::Response, <::abstract_adapter::mock::MockAdapterContract as ::abstract_sdk::base::Handler>::Error> {
+            use ::abstract_sdk::base::ExecuteEndpoint;
+            Ok(::cosmwasm_std::Response::new().set_data("mock_exec".as_bytes()))
+        }
+
+        /// Query entrypoint
+        fn query(
+            deps: ::cosmwasm_std::Deps,
+            env: ::cosmwasm_std::Env,
+            msg: <::abstract_adapter::mock::MockAdapterContract as ::abstract_sdk::base::QueryEndpoint>::QueryMsg,
+        ) -> Result<::cosmwasm_std::Binary, <::abstract_adapter::mock::MockAdapterContract as ::abstract_sdk::base::Handler>::Error> {
+            use ::abstract_sdk::base::QueryEndpoint;
+            MOCK_ADAPTER.query(deps, env, msg)
+        }
+
+        type Exec = ::abstract_core::base::ExecuteMsg<::abstract_core::adapter::AdapterBaseMsg, MockExecMsg, MockReceiveMsg>;
+        type Query = ::abstract_core::adapter::QueryMsg<MockQueryMsg>;
+        type Init = ::abstract_core::adapter::InstantiateMsg<MockInitMsg>;
+        #[cw_orch::interface(Init, Exec, Query, Empty)]
+        pub struct $name ;
+
+        impl ::abstract_interface::AdapterDeployer<::cw_orch::prelude::Mock, MockInitMsg> for $name <::cw_orch::prelude::Mock> {}
+
+        impl Uploadable for $name<::cw_orch::prelude::Mock> {
             fn wrapper(&self) -> <Mock as ::cw_orch::environment::TxHandler>::ContractSource {
                 Box::new(ContractWrapper::<
                     Exec,