+++ framework/packages/abstract-core/src/core/manager.rs
@@ -90,6 +90,8 @@ pub mod state {
     pub const SUB_ACCOUNTS: Map<u32, cosmwasm_std::Empty> = Map::new("sub_accs");
     /// Pending new governance
     pub const PENDING_GOVERNANCE: Item<GovernanceDetails<Addr>> = Item::new("pgov");
+    /// Context for old adapters that are currently removing authorized addresses
+    pub const REMOVE_ADAPTER_AUTHORIZED_CONTEXT: Item<u64> = Item::new("rm_a_auth");
 }
 
 use self::state::AccountInfo;