+++ framework/packages/abstract-interface/src/deployment.rs
@@ -76,8 +76,10 @@ impl<Chain: CwEnv> Deploy<Chain> for Abstract<Chain> {
 
         // Set Factory
         deployment.version_control.execute(
-            &abstract_core::version_control::ExecuteMsg::SetFactory {
-                new_factory: deployment.account_factory.address()?.into_string(),
+            &abstract_core::version_control::ExecuteMsg::UpdateConfig {
+                account_factory_address: Some(deployment.account_factory.address()?.into_string()),
+                namespace_registration_fee: None,
+                allow_direct_module_registration_and_updates: None,
             },
             None,
         )?;